+++ CHANGES.md
@@ -4,4 +4,10 @@ significant changes between each commit will be written below to avoid clogging 
 
 # changes:
 
-- removed credentials from index.js (accidentally committed them last time, they've been updated since)
+- added clientside TournamentManager for managing tournament scoring
+	- client can now request for match data from the server via websocket
+- added serverside ChallongeTwoStageTournament class for managing the challonge api with these tourneys specifically
+- added a bunch of ChallongeAPI functionality
+- added options.json to server folder for extra configuration
+- added more info to README
+- added more funny pictures (very important)